fix: backend 보안 CRITICAL 5건 + HIGH 5건 수정
CRITICAL:
- C1: 비밀번호 8자+복잡도 검증, role/team Literal 타입 제한
- C2: 로그인 rate limiting (slowapi + Redis, 5회/분)
- C3: HTTP 보안 헤더 6종 (CSP, X-Frame-Options 등)
- C4: 임시 비밀번호 HTTP 응답 노출 제거
- C5: 금액 필드 int → Decimal 통일 (contract, fund_lp)

HIGH:
- H3: partner RBAC bypass → compliance/cap_table 접근 차단
- H5: 사용자 목록 API admin 권한 체크 추가
- H6: DealStage 전환 화이트리스트 강제 (terminal 상태 거부)
- H7: cap_table investment_amount Integer → BigInteger
- H8: exit_record exit_amount Integer → BigInteger

Gap 분석 결과: 100% (31/31 PASS)

+++ backend/app/models/cap_table.py
@@ -4,7 +4,7 @@ import uuid
 from datetime import date, datetime
 from decimal import Decimal
 
-from sqlalchemy import Boolean, Date, ForeignKey, Integer, Numeric, String, Text, func
+from sqlalchemy import BigInteger, Boolean, Date, ForeignKey, Integer, Numeric, String, Text, func
 from sqlalchemy.orm import Mapped, mapped_column
 
 from app.models.base import Base
@@ -19,7 +19,7 @@ class CapTableEntry(Base):
     share_type: Mapped[str] = mapped_column(String(30))  # common / preferred / rcps / option
     shares: Mapped[int] = mapped_column(Integer)
     ownership_pct: Mapped[Decimal] = mapped_column(Numeric(10, 4))
-    investment_amount: Mapped[int | None] = mapped_column(Integer, nullable=True)
+    investment_amount: Mapped[int | None] = mapped_column(BigInteger, nullable=True)
     investment_date: Mapped[date | None] = mapped_column(Date, nullable=True)
     round_name: Mapped[str | None] = mapped_column(String(50), nullable=True)
     notes: Mapped[str | None] = mapped_column(Text, nullable=True)
